Conflicts:
	CS2103V01G30/bin/Debug/CS2103V01G30.exe
	CS2103V01G30/bin/Debug/CS2103V01G30.pdb
	CS2103V01G30/bin/Debug/venues.txt
	CS2103V01G30/bin/Debug/z_Task.txt
	CS2103V01G30/obj/x86/Debug/CS2103V01G30.exe
	CS2103V01G30/obj/x86/Debug/CS2103V01G30.pdb

+++ CS2103V01G30/Control.xaml.cs
@@ -43,7 +43,7 @@ namespace CS2103V01G30
         static List<Budget> myBudgetList = new List<Budget>();
 
         EventMgt eventMgt = new EventMgt();
-        budgetMgmt budgetMgt = new budgetMgmt();
+        BudgetMgmt budgetMgt = new BudgetMgmt();
         Users stu = new Users();
         RegEvent regEvt = new RegEvent();
         RegProManagement regProMgt = new RegProManagement();
@@ -75,7 +75,6 @@ namespace CS2103V01G30
             //create the path of budgets
             myBudgetsPath = System.IO.Path.Combine(matricNo, "budgets.txt");
 
-
             budgetNtf = new BudgetNotif(myBudgetsPath);
 
             regEvt.setRegEvent(myName, matricNo);
@@ -91,6 +90,7 @@ namespace CS2103V01G30
                 regProMgt.addToRegProMgtList(myEventList[i].getEventName());
             }
 
+            //initialize all the lists
             showlistViewAllEvent();
             showlistViewMyEvent();
             showlistBoxBudget();
@@ -98,14 +98,13 @@ namespace CS2103V01G30
             showlistBoxAllVenue();
             showlistBoxNotification();
 
+            //show the cuurent time
             ShowTimer = new DispatcherTimer();
-            ShowTimer.Tick += new EventHandler(ShowCurTimer);//use timer
+            ShowTimer.Tick += new EventHandler(ShowCurTimer);
             ShowTimer.Interval = new TimeSpan(0, 0, 0, 1, 0);
             ShowTimer.Start();
             buttonRegister.Visibility = Visibility.Hidden;
             buttonCancel.Visibility = Visibility.Hidden;
-
-
         }
 
         #region notification
@@ -167,12 +166,10 @@ namespace CS2103V01G30
                 string studentInfo="";
                 foreach (Users user in aUser.userlist)
                 {
-
                     studentInfo += "ID:" + user.username + " HP: " + user.contact + " Email: " + user.email + "\t";
                 }
                 listViewTask.Items.Add(new { Task = topTaskMgt.getTaskName(eventName, i), Deadline = topTaskMgt.getDueDate(eventName, i), Person = topTaskMgt.getPersonName(eventName, i), Status = topTaskMgt.getStatus(eventName, i), StudentInfo = studentInfo });
             }
-            //string s = myEventList[listViewMyEvent.SelectedIndex].getEventName();
 
             labelTotaltask.Content = "Total Task: " + topTaskMgt.getNoOfTasksInATaskList(eventName);
             labelFinished.Content = "Finished: " + topTaskMgt.getNoOfDoneTasksInATaskList(eventName);
@@ -185,7 +182,6 @@ namespace CS2103V01G30
         {
             topTaskMgt.addTaskInTaskList(myEventList[listViewMyEvent.SelectedIndex].getEventName());
             showlistViewTask();
-            // showTotalTask();
         }
 
         private void buttonDelTask_Click(object sender, RoutedEventArgs e)
@@ -194,7 +190,6 @@ namespace CS2103V01G30
             {
                 topTaskMgt.deleteTaskInTaskList(myEventList[listViewMyEvent.SelectedIndex].getEventName(), listViewTask.SelectedIndex);
                 showlistViewTask();
-                //showTotalTask();
             }
             catch
             {
@@ -233,7 +228,7 @@ namespace CS2103V01G30
 
             }
             catch { }
-            }
+        }
 
         private void buttonEdit_Click(object sender, RoutedEventArgs e)
         {
@@ -311,6 +306,7 @@ namespace CS2103V01G30
         {
             try
             {
+                //get event ID, name, starting and ending time
                 int id = 0;
                 string name = textBoxName1.Text;
 
@@ -358,9 +354,18 @@ namespace CS2103V01G30
                     return;
                 }
 
+                //get venue and description
                 string v = labelVenue.Content.ToString();
                 string des = textBoxDescription.Text;
 
+                //description cannot be empty
+                if (des == String.Empty)
+                {
+                    MessageBox.Show("Please do not leave the description empty.");
+                    return;
+                }
+
+                //check if the venue chosen has a clash with the current reservation status
                 foreach (Venue ven in venueMgt.venueList)
                 {
                     if (v == ven.getLocation())
@@ -378,8 +383,8 @@ namespace CS2103V01G30
                         }
                 }
 
+                //create the new event
                 id = eventMgt.createEvent(name, startDate, endDate, sTime, eTime, v, des);
-
                 if (id != 0)
                 {
                     stu.add_oneCreatedEvent(matricNo, id);
@@ -387,12 +392,14 @@ namespace CS2103V01G30
                     MessageBox.Show("A event has been successfully created.");
                 }
 
+                //update the data
                 eventMgt.writeToFile();
                 budgetMgt.createBudget();
                 budgetMgt.writeToFile();
                 regProMgt.addToRegProMgtList(name);
                 topTaskMgt.addATaskManagement(name);
 
+                //update all the related list
                 showlistViewMyEvent();
                 showlistViewAllEvent();
                 showlistBoxBudget();
@@ -415,6 +422,7 @@ namespace CS2103V01G30
                     return;
                 }
 
+                //get event ID, name, starting and ending time
                 int eventID = myEventList[index].getEventID();
                 string name = textBoxName1.Text;
 
@@ -459,21 +467,32 @@ namespace CS2103V01G30
                     return;
                 }
 
+                //get venue and description
                 string v = labelVenue.Content.ToString();
                 string des = textBoxDescription.Text;
 
+                //description cannot be empty
+                if (des == String.Empty)
+                {
+                    MessageBox.Show("Please do not leave the description empty.");
+                    return;
+                }
+
+                //delete the occupied dates
                 foreach (Venue ven in venueMgt.venueList)
                 {
                     if (v == ven.getLocation())
                     {
                         for (int date = myEventList[index].getStartDate(); date <= myEventList[index].getEndDate(); date++)
                             ven.deleteOccupiedDate(date);
+                        
                         venueMgt.writeToFile();
                         venueMgt.readFromFile();
                         break;
                     }
                 }
                 
+                //check if there is a clash on venue booking
                 foreach (Venue ven in venueMgt.venueList)
                 {
                     if (v == ven.getLocation())
@@ -486,7 +505,7 @@ namespace CS2103V01G30
                                 {
                                     for (int d = myEventList[index].getStartDate(); d < myEventList[index].getEndDate(); d++)
                                         ven.addOccupiedDate(d);
-                                    MessageBox.Show("The venue you chose has been occupied on " + date.ToString("d"));
+                                    MessageBox.Show("The venue you chose has been reserved on " + date.ToString("d"));
                                     return;
                                 }
                             }
@@ -499,6 +518,7 @@ namespace CS2103V01G30
                     {
                         if (eventMgt.editEvent(eventID, name, startDate, endDate, sTime, eTime, v, des))
                         {
+                            //change the event info
                             myEventList[index].setEventName(name);
                             myEventList[index].setStartDate(startDate);
                             myEventList[index].setEndDate(endDate);
@@ -525,6 +545,7 @@ namespace CS2103V01G30
                         }
                     }
                 }
+                //update the data
                 venueMgt.writeToFile();
                 eventMgt.writeToFile();
                 showlistViewMyEvent();
@@ -630,7 +651,6 @@ namespace CS2103V01G30
                     topTaskMgt.addATaskManagement(myEventList[listViewMyEvent.SelectedIndex].getEventName());
                     regProMgt.addToRegProMgtList(myEventList[listViewMyEvent.SelectedIndex].getEventName());
                     showlistBoxBudget();
-                    //showTotalTask();
                     showlistViewTask();
                     showlistViewRegPro();
 
@@ -650,6 +670,7 @@ namespace CS2103V01G30
                         BitmapImage bitmapImage = new BitmapImage();
                         bitmapImage.BeginInit();
                         bitmapImage.StreamSource = new FileStream(imagePath, FileMode.Open, FileAccess.Read);
+                        
                         //load the image now so we can immediately dispose of the stream
                         bitmapImage.CacheOption = BitmapCacheOption.OnLoad;
                         bitmapImage.EndInit();
@@ -681,12 +702,22 @@ namespace CS2103V01G30
                 if (index <= eventMgt.eventList.Count && index >= 0)
                 {
                     labelpName.Content = "Name:    " + eventMgt.eventList[index].getEventName();
+
                     int date = eventMgt.eventList[index].getStartDate();
                     int day = date / 1000000;
                     int month = (date / 10000) % 10;
                     int year = date % 10000;
-                    string dateString = day.ToString() + "/" + month.ToString() + "/" + year.ToString();
-                    labelpDate.Content = "Date:     " + dateString;
+                    DateTime startDate = new DateTime(year, month, day);
+                    datePickerStartDate.SelectedDate = startDate;
+
+                    date = eventMgt.eventList[index].getEndDate();
+                    day = date / 1000000;
+                    month = (date / 10000) % 10;
+                    year = date % 10000;
+                    DateTime endDate = new DateTime(year, month, day);
+                    datePickerEndDate.SelectedDate = endDate;
+
+                    labelpDate.Content = "Date:     " + startDate.ToString("d") + " to " + endDate.ToString("d");
                     labelpTime.Content = "Time:     " + eventMgt.eventList[index].getStartTime().ToString() + " to " + eventMgt.eventList[index].getEndTime().ToString();
                     labelpVenue.Content = "Venue:    " + eventMgt.eventList[index].getVenue();
                     textBlockDescription.Text = "Description:    " + System.Environment.NewLine + eventMgt.eventList[index].getDescription();
@@ -697,6 +728,7 @@ namespace CS2103V01G30
                         BitmapImage bitmapImage = new BitmapImage();
                         bitmapImage.BeginInit();
                         bitmapImage.StreamSource = new FileStream(imagePath, FileMode.Open, FileAccess.Read);
+                        
                         //load the image now so we can immediately dispose of the stream
                         bitmapImage.CacheOption = BitmapCacheOption.OnLoad;
                         bitmapImage.EndInit();
@@ -854,7 +886,6 @@ namespace CS2103V01G30
 
         void showCalendarAvailableDates(int venueIndex)
         {
-            venueMgt.readFromFile();
             calendarAvailableDates.BlackoutDates.Clear();
 
             foreach (int date in venueMgt.venueList[venueIndex].occupiedDates)
@@ -938,11 +969,19 @@ namespace CS2103V01G30
                 }
 
                 if (datePickerSearchVenueDate.SelectedDate != null)
-                {
+                {                  
                     int year = datePickerSearchVenueDate.SelectedDate.Value.Year;
                     int month = datePickerSearchVenueDate.SelectedDate.Value.Month;
                     int day = datePickerSearchVenueDate.SelectedDate.Value.Day;
                     int targetDate = 1000000 * day + 10000 * month + year;
+
+                    DateTime searchDate = new DateTime(year, month, day);
+                    
+                    if (DateTime.Compare(searchDate, DateTime.Now) < 0)
+                    {
+                        MessageBox.Show("The date you searched should not be earlier than current date.");
+                        return;
+                    }
 
                     for (int i = 0; i < venueMgt.venueList.Count; i++)
                     {
@@ -998,10 +1037,12 @@ namespace CS2103V01G30
                     double spent = myBudgetList[id].getBudgetSpent();
                     ava -= spent;
 
+                    //if there is a budget surplus, the color is green, otherwise it is red
                     if (ava >= 0)
                         labelAvaBudget.Foreground = Brushes.Green;
                     else
                         labelAvaBudget.Foreground = Brushes.Red;
+                    
                     labelAvaBudget.Content = "Budget available: $" + ava;
                     labelSpentBudget.Content = "Budget spent: $" + myBudgetList[id].getBudgetSpent();
 
@@ -1054,7 +1095,7 @@ namespace CS2103V01G30
         }
 
         void buttonDelBudgetItem_Click(object sender, RoutedEventArgs e)
-        {
+        {   
             try
             {
                 int id = myEventList[listViewMyEvent.SelectedIndex].getEventID();
@@ -1069,8 +1110,8 @@ namespace CS2103V01G30
                 MessageBox.Show("Please select an item of the budget.");
             }
         }
-         private void budgetItemListView_MouseDoubleClick(object sender, MouseButtonEventArgs e)
-        //private void budgetItemListView_SelectionChange(object sender, SelectionChangedEventArgs e)
+         
+        private void budgetItemListView_MouseDoubleClick(object sender, MouseButtonEventArgs e)
         {
             textBoxItemName.Clear();
             textBoxAmountSpent.Clear();
@@ -1196,7 +1237,6 @@ namespace CS2103V01G30
                     if (eventName == eve.getEventName())
                     {
                         regEvt.cancelRegistration(listViewAllEvent.SelectedItem.ToString());
-                        // stu.remove_oneJoinedEvent(matricNo, eve.getEventID());
                         break;
                     }
                 }
@@ -1211,7 +1251,6 @@ namespace CS2103V01G30
 
         private void buttonApprove2_Click(object sender, RoutedEventArgs e)
         {
-
             try
             {
                 {
@@ -1285,6 +1324,7 @@ namespace CS2103V01G30
                     {
                         int CursorIndex = T.SelectionStart - 1;
                         T.Text = T.Text.Remove(CursorIndex, 1);
+                        
                         //Align Cursor to same index
                         T.SelectionStart = CursorIndex;
                         T.SelectionLength = 0;
@@ -1350,18 +1390,14 @@ namespace CS2103V01G30
             timer1.Text += "    ";
             //time
             timer1.Text += "Current time is: " + DateTime.Now.ToString("HH:mm:ss");
-            //System.Diagnostics.Debug.Print("this.ShowCurrentTime {0}", this.ShowCurrentTime);
         }
 
         private void UpdateInfo_Click(object sender, RoutedEventArgs e)
         {
             var updatePersonalInfo = new UpdateInfo();
             updatePersonalInfo.Show();
-
         }
-
         #endregion
-
     }
 }
